Move CampaignReferralRow
